src/backend/myapi/serializers.py
from django.contrib.auth.models import Group, User
from myapi.models import Item
from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['url', 'username', 'email', 'groups']


class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = ['url', 'name']


class ItemSerializer(serializers.Serializer):
    id = serializers.IntegerField(read_only=True)